+++ lib/ui/emulator/debugger/breakpoint_dialog.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nesd/extension/hex_extension.dart';
+import 'package:nesd/nes/debugger/breakpoint.dart';
 import 'package:nesd/nes/debugger/debugger.dart';
 import 'package:nesd/nes/debugger/debugger_state.dart';
-import 'package:nesd/nes/nes.dart';
 import 'package:nesd/ui/emulator/debugger/debugger_widget.dart';
 
 class BreakpointDialog extends ConsumerWidget {
@@ -23,17 +23,20 @@ class BreakpointDialog extends ConsumerWidget {
 
     return AlertDialog(
       title: const Text('Breakpoints'),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          for (final breakpoint in debuggerState.breakpoints)
-            if (!breakpoint.hidden)
-              BreakpointRow(
-                breakpoint: breakpoint,
-                scrollController: scrollController,
-              ),
-          const AddBreakpointWidget(),
-        ],
+      content: SizedBox(
+        width: 300,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            for (final breakpoint in debuggerState.breakpoints)
+              if (!breakpoint.hidden)
+                BreakpointRow(
+                  breakpoint: breakpoint,
+                  scrollController: scrollController,
+                ),
+            const AddBreakpointWidget(),
+          ],
+        ),
       ),
       actions: [
         TextButton(
@@ -62,31 +65,74 @@ class BreakpointRow extends ConsumerWidget {
     final debugger = ref.watch(debuggerProvider);
     final state = ref.watch(debuggerNotifierProvider);
 
-    return Row(
+    return Column(
       children: [
-        Text(
-          breakpoint.address.toHex(width: 4),
-          style: monoStyle.copyWith(fontSize: 15),
-        ),
-        const Spacer(),
-        IconButton(
-          icon: const Icon(Icons.gps_fixed),
-          tooltip: 'Go to address',
-          onPressed: () {
-            final pcOffset = calculateAddressScrollOffset(
-              state,
-              breakpoint.address,
-            );
+        Row(
+          children: [
+            Checkbox(
+              value: breakpoint.enabled,
+              onChanged: (_) =>
+                  debugger.toggleBreakpointEnabled(breakpoint.address),
+            ),
+            const SizedBox(width: 8),
+            Text(
+              breakpoint.address.toHex(width: 4),
+              style: monoStyle.copyWith(
+                fontSize: 15,
+                color: breakpoint.enabled ? null : Colors.grey,
+              ),
+            ),
+            const Spacer(),
+            IconButton(
+              icon: const Icon(Icons.gps_fixed),
+              tooltip: 'Go to address',
+              onPressed: () {
+                final pcOffset = calculateAddressScrollOffset(
+                  state,
+                  breakpoint.address,
+                );
 
-            jumpTo(scrollController, pcOffset);
+                jumpTo(scrollController, pcOffset);
 
-            Navigator.of(context).pop();
-          },
+                Navigator.of(context).pop();
+              },
+            ),
+            IconButton(
+              icon: const Icon(Icons.close),
+              tooltip: 'Remove breakpoint',
+              onPressed: () => debugger.removeBreakpoint(breakpoint),
+            ),
+          ],
         ),
-        IconButton(
-          icon: const Icon(Icons.close),
-          tooltip: 'Remove breakpoint',
-          onPressed: () => debugger.removeBreakpoint(breakpoint),
+        Row(
+          children: [
+            const SizedBox(width: 32),
+            Checkbox(
+              value: breakpoint.disableOnHit,
+              onChanged: (_) {
+                breakpoint.disableOnHit = !breakpoint.disableOnHit;
+
+                debugger.updateBreakpoint(breakpoint);
+              },
+            ),
+            const SizedBox(width: 4),
+            const Text('Disable on hit', style: TextStyle(fontSize: 15)),
+          ],
+        ),
+        Row(
+          children: [
+            const SizedBox(width: 32),
+            Checkbox(
+              value: breakpoint.removeOnHit,
+              onChanged: (_) {
+                breakpoint.removeOnHit = !breakpoint.removeOnHit;
+
+                debugger.updateBreakpoint(breakpoint);
+              },
+            ),
+            const SizedBox(width: 4),
+            const Text('Remove on hit', style: TextStyle(fontSize: 15)),
+          ],
         ),
       ],
     );